inquiry: create a dummy class to wrap the VPD numbers.
Create a dummy class, VPD, with one element for each supported
VPD value.
This is kind of an enum, but more portable since proper enums did not come
until python 3.4.

This changes the use slightly since you need to write
INQUIRY.VPD.SUPPORTED_VPD_PAGES instead of just
INQUIRY.SUPPORTED_VPD_PAGES.

The extra .VPD. is good for readability since it improves the information:
This is part of the INQUIRY command,
this is a VPD page constant
the actual page is SupportedVPDPages.
I.e. it hightlights that this is a VPD page constant.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>