test-inquiry.py: change the inquiry printout headers
Change the printout for the inquiry headers in the test tool.
The basic bage where EVPD==0 is called the standard data and
the extended format for PAGE_CODE==0 is called the
SUPPORTED_VPD_PAGES.

Later we will need to add other page_codes as well. Both SPC and
SBC contains lots of additional page_codes that we will need at some stage.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>