enum.py: looking up the value should return the key, not the value
When we lookup the value of an enum, return the string for the key
and not just return the value back (in possibly a different format).

If no key was found for the requested value then return "" since this
function is supposed to return a string. (.i.e. returning None would probably
not do what we want here.)

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>