scsi_device.py: use linux_sgio module instead of the sgio module
We are going to remove the builtin sgio module and instead use an external,
conditionally imported, module called linux_sgio.
Name is meant to indicate that this is a linux specific module.

This linux_sgio module is linux specific and will not be available on
non-linux systems which is why we import it conditionally but will not
fail if it is not available.
This is to allow using python-scsi on other, non-linux, systems.
The aim is to allow to create windows/solaris/ios/... specific modules for
accessing scsi devices and plugging them into python-scsi just like the
existing libiscsi module we already have and the linux_sgio module we add in
this patch.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>