linux_sgio: add an external module for linux_sgio devices
This is mainly shifting code around. This is basically the existing
sgio/sgiomodule.c copied into the linux_sgio/ directory. This is part of the
steps to remove the currently builtin module for sgio which only can be used
under linux and to make that same module a separate part to allow using
python-scsi also on non-linux platforms.

Later patches will remove the sgio/sgiomodule.c module completely.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>