pyscsi: rename the main module from sgio to pyscsi
Rename the main module from sgio to pyscsi to better reflect that this
is a python module for doing SCSI.
The previous name was poor since it hinted at that this module was specific
to SGIO, which is no longer the case. The current module is specific to SCSI
and SGIO is no just one, of several, modules that can be used to talk to an
actual device.

This now means we have a directory structure of
pyscsi/pyscsi
and
pyscsi/utils

We should later probably change this to just all reside in pyscsi instead.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>