SCSI: add cdb marshalling/unmarshalling be dict driven and datain marshalling
ReadCapacity10/16

Add a helper function encode_dict that is the inverse of decode_bits.
This function expects the same dict containing the schema as decode_bits
and takes a dict containing values as input marshalls this dict into a
bytearray. This is the opposite of decode_bits which takes a bytearray
as input and unmarshalls it into a dict.

Add static methods for cdb marshalling and unmarshalling the cdb.
Make cdb marshalling dict driven instead of manual marshalling of the fields
we used before.

Add static methods for marshalling and unmarshalling of the data-in buffer
for these commands.

Add tests to verify marshalling and test that
<dict> == unmarshall_datain(marshall_datain(<dict>))

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>