modesense6: add datain unmarshalling and make it all dict driven
Return a set of mode pages under 'mode_pages' in the dict instead of only
a single page. We will need this in order to support "report all mode pages"
and friends.
This means we have to do a slight change to the mtx tool which inspects the
modepage since it has now moved around a little in the dict.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>

Conflicts:
	pyscsi/pyscsi/scsi_cdb_modesense6.py
	tests/test_unmarshall_modesense6.py