ModeSelect6: Add ModeSelect6 support
Add support for ModeSelect6

ModeSelect6 is implemented as a separate class in the same file as used for
ModeSense6 and not as a new file. This for two main reasons.

ModeSelect6 shares almost all its structure definitions with ModeSense6
and also, it is unlikely that anyone will ever use ModeSelect6 by itself.
Instead the expected workflow is that a user will use ModeSense6 to fetch a
mode page from the target, then mutate the dict slightly and send the mutated
dict back to the target again.

For the second case it would be very annoying to have to import two different
files instead of just importing modesense6 and then also automagically get all
the definitions for modeselect6.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>