fixing encoding and decoding in the test_unmarshall_inquiry
with python 3 we need to do some explicit encoding and decoding for
strings and characters. so bytearray needs to get the int value of a character
and to check for the string we need to decode the bytearray to unicode string.
The same we have to do if we pass a string to the bytearray, we need to
encode it. this code runs under python 2 also so noting should be broken.