remove  SCSI dependency from SCSICommand
the SCSICommand takes a OpCode instead a SCSI instance now. All the hidden execute
functionality is done outside the SCSICommand class. SCSI Class has now two wrapper
to make it a bit more convenient to call execute on the SCSIDevice and the unmarshall
method on SCSICommand. The wrapper methods in the SCSI class do all the work
removed from the SCSICommand execute method now.