restructure the code
for the sake of readability the function and method args
going each on a new line. This is done in the definition and
at the call of the method or function.

Also align dicts  even we have a lot of
whitespaces after the equal sign. This doesn't apply to the
enum modules because there every section is readable still.

If someone feel the urge  to re-PEP8 it ...don't, I done it before
and now I undo it again!

Signed-off-by: Markus Rosjat <markus.rosjat@gmail.com>