make unmarshall_cdb and marshall_cdb a static method again
also build the cdb dict in build_cdb from kwargs. This will cause
a passing of kwargs in the __init__ method of SCSICommand subclasses.

unmarshall is a method of SCSICommand now. The wrapper method
in SCSI class will be removed. There need to be a proper checking
if the Subclass  implements  a unmarshall_datain method ...still.

Signed-off-by: Markus Rosjat <markus.rosjat@gmail.com>