initial commit for a simple version of libiscsi-python
we just need a few functions to be created by SWIG so we just define
a interface file suited for pyscsi. he mthere is no genarated c file
since we are going to run this against a local libiscsi installation.

the Makefile will just generate a c file, building the extention is
build by setup.py for now

Signed-off-by: Markus Rosjat <markus.rosjat@gmail.com>