Remove linux_sgio extension in favour of the cython-sgio module.
As I only need the minimal interface provided by the Cython wrapper I added
to this set of changes in one of my projects, I thought it would be faster
to just factor it out, and depend on it here instead.

This removes some of the complexity in this package, and allows independent
releases of the Python code wrapping the SCSI commands, and the Cython code
sending the ioctl.