Replace bundled libiscsi and sgio modules with external Cython ones.
This uses the modules as provided in:

 - https://github.com/python-scsi/cython-iscsi
 - https://github.com/python-scsi/cython-sgio

which can be installed separately, removing the dependency on SWIG
altogether.

Note that trying to keep both these and the new Cython modules in the same
repository is more complicated than it appear worth, as the new modules
have a slight, but significant, difference in their APIs.