Make tests runnable without installing the package.
Declare `tests` a package (by creating `__init__.py`), which allows
the imports to `pyscsi` to resolve even if it's not installed (the
installed copy take precedence).

For this to work, the `mock_device` module needs to now be imported
relatively rather than absolutely (`from .mock_device`), which does
not work when executing the tests manually (and thus via the `make`
command).

Results:

 - running tests is much easier, when checking out the repository to
   make simple changes;
 - running tests is no longer dependent on Unix-like platforms;
 - to run a single test file, you can use `pytest tests/test_enum.py`
   instead.