Add `.gitignore` files to make it much easier to contribute from Windows.
Without a `.gitignore` in place, most of the Windows integration for Git
(GitHub Desktop, Visual Studio Code) assumes you want to commit _all_ of
the new files — or suggests adding them to `.gitignore`.

Given that `.travis.yml` is also repository (but not code) configuration,
this should now be an acceptable compromise.